# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Arc Pass API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get the current user profile, including connected accounts
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/wallets:
    get:
      operationId: listUserWallets
      tags: [users]
      summary: List wallets linked to the current user
      responses:
        "200":
          description: Wallet list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallets/challenge:
    post:
      operationId: createWalletChallenge
      tags: [wallets]
      summary: Create an expiring wallet ownership challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletChallengeInput"
      responses:
        "201":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletChallenge"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallets/verify:
    post:
      operationId: verifyWalletOwnership
      tags: [wallets]
      summary: Verify a wallet signature and save the ownership-verified wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletVerificationInput"
      responses:
        "201":
          description: Wallet ownership verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /users/me/wallets/{walletId}:
    delete:
      operationId: removeUserWallet
      tags: [users]
      summary: Remove a wallet from the current user
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Wallet removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /share/{type}/{id}:
    get:
      operationId: getSharePage
      tags: [sharing]
      summary: Public share page with Open Graph metadata
      parameters:
        - { name: type, in: path, required: true, schema: { type: string, enum: [founder, builder] } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: HTML share page
          content:
            text/html:
              schema: { type: string }
  /share/{type}/{id}/image:
    get:
      operationId: getShareImage
      tags: [sharing]
      summary: Server-generated Open Graph image
      parameters:
        - { name: type, in: path, required: true, schema: { type: string, enum: [founder, builder] } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: SVG Open Graph image
          content:
            image/svg+xml:
              schema: { type: string, format: binary }
  /webhooks/typeform/founder-application:
    post:
      operationId: ingestFounderApplication
      tags: [webhooks]
      summary: Ingest a signature-verified Typeform founder application
      responses:
        "204": { description: Accepted or previously processed }
        "401":
          description: Invalid webhook signature
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserConnections:
      type: object
      required: [x, discord, github]
      properties:
        x:
          type: object
          required: [connected]
          properties:
            connected:
              type: boolean
            username:
              type: ["string", "null"]
        discord:
          type: object
          required: [connected]
          properties:
            connected:
              type: boolean
            username:
              type: ["string", "null"]
            avatarUrl:
              type: ["string", "null"]
        github:
          type: object
          required: [connected]
          properties:
            connected:
              type: boolean
            username:
              type: ["string", "null"]
    UserProfile:
      type: object
      required: [id, username, displayName, avatarUrl, provider, isAdmin, createdAt, connections]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        provider:
          type: string
          enum: [x, discord]
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
        connections:
          $ref: "#/components/schemas/UserConnections"
    Wallet:
      type: object
      required: [id, address, isPrimary, label, chain, ownershipVerifiedAt, signatureMethod, createdAt]
      properties:
        id:
          type: integer
        address:
          type: string
        isPrimary:
          type: boolean
        label:
          type: ["string", "null"]
        chain:
          type: string
        ownershipVerifiedAt:
          type: string
          format: date-time
        signatureMethod:
          type: string
        lastAnalysedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    WalletChallengeInput:
      type: object
      required: [address]
      properties:
        address:
          type: string
    WalletChallenge:
      type: object
      required: [challengeId, address, message, expiresAt]
      properties:
        challengeId:
          type: integer
        address:
          type: string
        message:
          type: string
        expiresAt:
          type: string
          format: date-time
    WalletVerificationInput:
      type: object
      required: [challengeId, message, signature]
      properties:
        challengeId:
          type: integer
        message:
          type: string
        signature:
          type: string
        label:
          type: string
        isPrimary:
          type: boolean
...